In window_filter, allow ignoring apps by bundleID
Created a table hs.window.filter.bundleIDignoreAlways to
enumerate bundleIDs that should be ignored. Initialize it with
com.apple. bundles.

Added isGuiByBundleID(bundleid) that returns false if
the given bundleID is in this list.

Its implementation is almost identical to isGui(appName) that
returns false is the application is listed in
hs.window.filter.ignoreAlways